+++ projects/pioneer-charts-docs/src/app/app.routes.ts
@@ -3,6 +3,7 @@ import { Routes } from '@angular/router';
 export const routes: Routes = [
     { path: 'docs/guides/introduction', loadComponent: () => import('./docs/introduction/introduction.component').then(m => m.IntroductionComponent) },
     { path: 'docs/guides/theme', loadComponent: () => import('./docs/theme/theme.component').then(m => m.ThemeComponent) },
+    { path: 'docs/guides/height-full', loadComponent: () => import('./docs/height-full/height-full.component').then(m => m.HeightFullComponent) },
     { path: 'docs/guides/data-contract', loadComponent: () => import('./docs/data-contract/data-contract.component').then(m => m.DataContractComponent) },
     { path: 'docs/components/charts/legend', loadComponent: () => import('./docs/legend/legend.component').then(m => m.LegendComponent) },
     { path: 'docs/components/charts/bar-chart', loadComponent: () => import('./docs/bar-chart/bar-chart.component').then(m => m.BarChartComponent) },